+++ Libraries/PeriphDrivers/Source/TMR/tmr_revc.c
@@ -260,7 +260,7 @@ void MXC_TMR_RevC_SetCount(mxc_tmr_regs_t *tmr, uint32_t cnt)
     tmr->cnt = cnt;
 }
 
-void MXC_TMR_RevC_TO_Start(mxc_tmr_regs_t *tmr, unsigned long us)
+void MXC_TMR_RevC_TO_Start(mxc_tmr_regs_t *tmr, uint32_t us)
 {
     uint64_t ticks;
     int clk_shift = 0;
